Configura projeto para trabalhar com Postgres

+++ Einstein.Core/Models/Aluno.cs
@@ -1,16 +1,16 @@
 ﻿namespace Einstein.Core.Models
 {
     public class Aluno : PessoaFisica
-    {                          
+    {
         public int? AnoConclusaoEnsinoMedio { get; set; }
         public bool JaRealizouEnem { get; set; }
-        public double? MediaGeralEnem { get; set; }
-        public double? NotaRedacaoEnem { get; set; }
-        public double? NotaExatasEnem { get; set; }
-        public double? NotaCienciasDaNaturezaEnem { get; set; }
-        public double? NotaHumanasEnem { get; set; }
-        public double? NotaLinguagensEnem { get; set; }                        
-        public bool EstaEmpregado { get; set; }                
-        public bool ExAluno { get; set; }
+        public float? MediaGeralEnem { get; set; }
+        public float? NotaRedacaoEnem { get; set; }
+        public float? NotaExatasEnem { get; set; }
+        public float? NotaCienciasDaNaturezaEnem { get; set; }
+        public float? NotaHumanasEnem { get; set; }
+        public float? NotaLinguagensEnem { get; set; }
+        public bool EstaEmpregado { get; set; }
+        public bool ExAluno { get; set; }        
     }
 }
